Check if possible to respond to the load added & round response value to 0.1

+++ gem-spaas.powerplant.coding-challenge/ViewModel/PayloadVm.cs
@@ -8,7 +8,7 @@ namespace gem_spaas.powerplant.coding_challenge.ViewModel
 {
     public class PayloadVm
     {
-        public int Load { get; set; }
+        public double Load { get; set; }
         public Fuel Fuels { get; set; }
         public List<Powerplant> Powerplants { get; set; }
     }
